+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -3,8 +3,8 @@ package roomescape.service;
 import org.springframework.stereotype.Service;
 import roomescape.domain.Reservation;
 import roomescape.domain.ReservationTime;
-import roomescape.dto.reservationtime.ReservationTimeCreateRequest;
-import roomescape.dto.reservationtime.ReservationTimeResponse;
+import roomescape.service.dto.reservationtime.ReservationTimeCreateRequest;
+import roomescape.service.dto.reservationtime.ReservationTimeResponse;
 import roomescape.exception.BadRequestException;
 import roomescape.exception.ResourceNotFoundException;
 import roomescape.repository.reservation.ReservationRepository;
