+++ src/main/java/roomescape/service/dto/reservationtime/ReservationTimeCreateRequest.java
@@ -1,4 +1,4 @@
-package roomescape.dto.reservationtime;
+package roomescape.service.dto.reservationtime;
 
 import roomescape.domain.ReservationTime;
