+++ src/test/java/roomescape/service/ReservationTimeServiceTest.java
@@ -10,8 +10,8 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import roomescape.Fixtures;
 import roomescape.domain.Reservation;
 import roomescape.domain.ReservationTime;
-import roomescape.dto.reservationtime.ReservationTimeCreateRequest;
-import roomescape.dto.reservationtime.ReservationTimeResponse;
+import roomescape.service.dto.reservationtime.ReservationTimeCreateRequest;
+import roomescape.service.dto.reservationtime.ReservationTimeResponse;
 import roomescape.exception.BadRequestException;
 import roomescape.repository.reservation.ReservationRepository;
 import roomescape.repository.reservationtime.ReservationTimeRepository;
